Collect hints in lychee-lib instead of lychee-bin
