+++ lychee-lib/src/lib.rs
@@ -102,7 +102,7 @@ pub use crate::{
         BaseInfo, BasicAuthCredentials, BasicAuthSelector, CacheStatus, CookieJar, ErrorKind,
         FileExtensions, FileType, Input, InputContent, InputResolver, InputSource, LycheeResult,
         Preprocessor, Redirect, Redirects, Request, RequestError, ResolvedInputSource, Response,
-        ResponseBody, Result, Status, StatusCodeSelector, StatusRange, StatusRangeError,
+        ResponseBody, Result, Status, StatusCodeSelector, StatusRange, StatusRangeError, hints::*,
         uri::raw::RawUri, uri::raw::RawUriSpan, uri::valid::Uri,
     },
 };
